Color Palette/Number controls: trimmed the fat
- decreased size/distance between controls for Color Palette window
- set max width of number controls to be maxBytes * stringwidth of a character
- allow sliders to be compressed horizontally
- change hex value to lowercase
- re-add "shift" modifier to align sliders for only RGB

+++ artpaint/controls/HSVColorControl.cpp
@@ -89,7 +89,7 @@ HSVColorControl::SetSliderColors(rgb_color c)
  	slider1->Slider()->SetColors(&colorList);
 
  	while (colorList.CountItems() > 0) {
- 		uint32* hue_word = (uint32*)(colorList.RemoveItem(0));
+ 		uint32* hue_word = (uint32*)(colorList.RemoveItem((int32)0));
  		delete hue_word;
  	}
